+++ client/game/Empire/store/index.js
@@ -40,7 +40,7 @@ const store = new Vuex.Store({
         stars: state => state.stars,
         player: state => state.player,
         playerResources: state => state.resources,
-        getPlanetById: state => id => state.planets.find(planet => planet.id === id) || {},
+        planetById: state => id => state.planets.find(planet => planet.id === id) || {},
         harvesterById: state => id => state.harvesters.find( harvester => harvester.id === id) || {},
         pdusByPlanetId: state => id => state.pdus.filter(pdu => pdu.planet === id) || []
     },