+++ server/handlers/template.js
@@ -12,7 +12,7 @@ const i18n = require("i18n"); // https://github.com/mashpie/i18n-node
 const cfg = require("../config");
 
 // moment.js is a handy library for displaying dates. We need this in our templates to display things like "Posted 5 minutes ago"
-exports.moment = require( "moment" );
+exports.moment = require("moment");
 
 // Dump is a handy debugging function we can use to sort of "console.log" our data
 exports.dump = obj => JSON.stringify(obj, null, 2);
@@ -21,8 +21,7 @@ exports.dump = obj => JSON.stringify(obj, null, 2);
 //exports.staticMap = ([lng, lat]) => `https://maps.googleapis.com/maps/api/staticmap?center=${lat},${lng}&zoom=14&size=800x150&key=${process.env.MAP_KEY}&markers=${lat},${lng}&scale=2`;
 
 // inserting a SVG image
-exports.image = path =>
-    fs.readFileSync(`./server/public/assets/images/${path}`);
+exports.image = path => fs.readFileSync(`./server/public/assets/images/${path}`);
 
 /*
  * inserting a SVG icon with <use>
@@ -36,22 +35,28 @@ exports.icon = (name, modifiers) => {
     modifiers = modifiers.map(modifier => {
         return "aw-icon--" + modifier;
     });
-    return `<svg class="aw-icon ${modifiers.join(
-        " "
-    )}"><use xlink:href="#${name}"></use></svg>`;
+    return `<svg class="aw-icon ${modifiers.join(" ")}"><use xlink:href="#${name}"></use></svg>`;
 };
 
 /*
  * insert inline spritesheet into body.
  * this means we don't use caching for this, but we avoid a lot of problems.
  */
+exports.gameIcons = () =>
+    fs
+        .readFileSync("./server/public/assets/images/icons-game.svg", {encoding: "utf-8"})
+        .replace("<svg", `<svg class="aw-icon-sprite"`)
+        .replace(`<?xml version="1.0" encoding="UTF-8"?>`, "");
+
+/*
+ * insert inline spritesheet of game icons for routes that include "game".
+ * this means we don't use caching for this, but we avoid a lot of problems.
+ */
 exports.spritesheet = () =>
     fs
-        .readFileSync(`./server/public/assets/images/icons-app.svg`, {
-            encoding: "utf-8"
-        })
-        .replace("<svg", '<svg class="aw-icon-sprite"')
-        .replace('<?xml version="1.0" encoding="UTF-8"?>', "");
+        .readFileSync(`./server/public/assets/images/icons-app.svg`, {encoding: "utf-8"})
+        .replace("<svg", `<svg class="aw-icon-sprite"`)
+        .replace(`<?xml version="1.0" encoding="UTF-8"?>`, "");
 
 /*
  * progressbar helper function
@@ -84,8 +89,6 @@ exports.appName = cfg.app.title;
 exports.__ = (key, data) => {
     return i18n.__(key, data);
 };
-
-
 
 exports.getAvatar = (avatar, size) => {
     if (avatar) {